+++ .phpunit.result.cache
@@ -1,1 +1,1 @@
-{"version":1,"defects":{"FaturamentoModelTest::testGetTotalVendido":8,"FaturamentoModelTest::testGetTotalProdutosVendidos":8,"FaturamentoModelTest::testGetTopProdutosVendidos":8,"AtualizarCadastroModelTest::testVerifyPasswordById":8,"LoginModelTest::testFindUserByEmailAndPassword":7,"CadastroTeste::testCheckEmailExistsTrue":8,"CadastroTeste::testCheckEmailExistsFalse":8},"times":{"FaturamentoModelTest::testGetTotalVendido":0.014,"FaturamentoModelTest::testGetTotalProdutosVendidos":0,"FaturamentoModelTest::testGetTopProdutosVendidos":0,"AtualizarCadastroModelTest::testGetUserById":0.026,"AtualizarCadastroModelTest::testUpdateUserCategory":0.002,"AtualizarCadastroModelTest::testVerifyPasswordById":0.104,"LoginModelTest::testFindUserByEmailAndPassword":0.02,"LoginModelTest::testFindUserByEmailAndPasswordInvalid":0.001,"CadastroTeste::testCreateUserSuccess":0.01,"CadastroTeste::testCreateUserFailure":0,"CadastroTeste::testCheckEmailExistsTrue":0.036,"CadastroTeste::testCheckEmailExistsFalse":0}}
+{"version":1,"defects":{"FaturamentoModelTest::testGetTotalVendido":8,"FaturamentoModelTest::testGetTotalProdutosVendidos":8,"FaturamentoModelTest::testGetTopProdutosVendidos":8,"AtualizarCadastroModelTest::testVerifyPasswordById":8,"LoginModelTest::testFindUserByEmailAndPassword":7,"CadastroTeste::testCheckEmailExistsTrue":8,"CadastroTeste::testCheckEmailExistsFalse":8,"CadastroModelTest::testCheckEmailExistsTrue":8,"CadastroModelTest::testCheckEmailExistsFalse":8},"times":{"FaturamentoModelTest::testGetTotalVendido":0.014,"FaturamentoModelTest::testGetTotalProdutosVendidos":0,"FaturamentoModelTest::testGetTopProdutosVendidos":0,"AtualizarCadastroModelTest::testGetUserById":0.026,"AtualizarCadastroModelTest::testUpdateUserCategory":0.002,"AtualizarCadastroModelTest::testVerifyPasswordById":0.104,"LoginModelTest::testFindUserByEmailAndPassword":0.02,"LoginModelTest::testFindUserByEmailAndPasswordInvalid":0.001,"CadastroTeste::testCreateUserSuccess":0.01,"CadastroTeste::testCreateUserFailure":0,"CadastroTeste::testCheckEmailExistsTrue":0.036,"CadastroTeste::testCheckEmailExistsFalse":0,"CadastroModelTest::testCheckEmailExistsTrue":0.033,"CadastroModelTest::testCheckEmailExistsFalse":0}}
